Fix: Dropping a piece clears the board (usually)
This is a fun one. I thought it must've been some race condition, but nah. When
starting the game by clicking the <Start> button, the button gets focus.
Subsequent press on the <Spacebar> key on the keyboard clicks the button
*facepalm*. Because of this, it manifests more towards the beginning of the
game, because as the game progresses, it is more likely that I'd have clicked
away from the <Start> button, causing it to lose focus (especially seeing as
I'm debugging the code, not merely playing for enjoyment.

The fix is straightforward, do not allow the button to keep focus.