Fix game over check; do not add extra piece when game over
This commit fixes the game over check to correctly determine if the game is over and prevents a new piece from being created if the game is over, or is about to be over.

This fixes the bug when at the end of the game, an extra piece is inserted and it partly overlaps the last piece, which happens almost always.