Responsive CSS
Convert all hard-coded pixel values (with the notable exception of border-radius, which just looks better constant) to a fraction of the viewport height (for horizontal displays) and viewport width (for vertical displays) respectively.

The media query to decide what to use as the basis of the calculation just simply cuts off at 600px display width, which probably isn't super-ideal and could be improved upon.