Rename button to "New Game" and fix button widths
The name 'Restart' for a button is just dorky; 'Start' was barely passable, but 'Restart' is truly unacceptable. 'New Game' is inviting and awesome.

Widen the layout column that houses the buttons; we will need that space for a preview window anyway. With the fluid layout it is quite easy to resize everything, took ten seconds.