fixes: gap width between board and buttons is wrong
This is a partial, no good, very bad, kludgy hack.

The problem is that the game-board as well as the overlays are *set* to have the same width (10 grid squares) as the game-container, but they *do not*. I have tried for about two hours to debug this, but have no clue what is actually going on, and setting padding/margin is also not useful. So I have created a new gutter in HTML and give it the width of the difference for low screen widths. It is still not ideal, but at least the layout doesn't look like arse.