Convert Autopilot methods to async, add model file placeholder, improve UI, refactor
This changes the Autopilot class in autopilot.js by:

* converting the maybeLoadScript method to an async function
* converting the loadModel method to an async function
* calling callback after the model is loaded in the maybeLoadScript method

It also adds a new file, ml/autopilot-model.json.

In script.js, it changes the TranslateTouchEventsToSyntheticKeyboardEvents function to be called Touch and adds an event listener for DOMContentLoaded to bind the Touch function to the gameBoard element.

In style.css, it adds some rules to reduce UI flicker for the #automate-button element.

In touch.js, it renames the TranslateTouchEventsToSyntheticKeyboardEvents function to Touch and adds an event listener for DOMContentLoaded to bind the Touch function to the gameBoard element when the DOM is ready, to reduce UI flicker.