Change the abstraction from keyboard move to position move
Also error handling is done just in a way that it all compiles and works; need to pour over the code and remove instrumentation once the bugs are fixed. All the print() statements are just instrumentation, it does nothing useful; same for console.log().

i.e., after watching Greer Viau's video "Building an AI to MASTER Tetris" https://www.youtube.com/watch?v=1yXBNKubb2o
I have, with the help of Github Copilot mostly, implemented a 40-state move in move.py and autopilot.js

This allows the AI to work not in terms of keyboard input, but in terms of where it wants the piece to end up, and in what rotation (4 rotations x 10 columns = 40; in reality the possible column positions will be 10 - piece's width in the respective rotation, and because most pieces are rotationally symmetric, some rotations will be identical)

Also replace the model with a dummy model, so that the whole thing compiles and we can test if all the pieces work together. We will improve the model later.

Because we're just using a dummy model, choose the best move according to the Reward.get_rewards() function