Rename the control class to Control
Long time ago when nothing worked yet, we had about seven ways to refer to the control class and its instance, depending on whether it was the class itself, an instance, a module, a `with` object, et cetera, that on top of the names for the Control class in JavaScript and its instances. Which is kind of funny, because the control class is a singleton. At some point during the debugging, I over-eagerly searched-and-replaced s/Control/control/, which was a mistake. But it wasn't more broken than before, in fact it seemed less broken, so I didn't fix it, and it stuck until now.