Train the model and enable it in the browser (finally everything works)
The model has trained for several hundred thousand iterations (the goal is 1M;
the trainer keeps getting killed after a few tens of thousands, so lost total
count), and now it mimics the reward function fairly well. It achieves about
1/4 the score the reward function does.

There was a compatibility issue between the tensorflow and tensorflowjs python
packages, so I downgraded the tensorflowjs package iteratively until it
successfully loaded.

The browser now actually loads and run the model.